feat: login and getting user info api calls
